// src/constants/types/assignment.types.ts
export type AssignmentDetailsType = {
  id: number;
  assignmentId: string;
  driverId: string;
  isAssigned: boolean;
  vecicle: string;
  requestedTo: string;
  assignmentDate: Date;
  startDate: Date;
  startTime: string;
  endDate: Date;
  endTime: string;
  createdAt: Date;
  updatedAt: Date;
};